requirements.txt
# packages in environment at /Users/marclawson/opt/anaconda3/envs/char_class:
#
# Name                    Version                   Build  Channel
appnope                   0.1.0                 py38_1001  
backcall                  0.2.0                      py_0  
blas                      1.0                         mkl  
boto3                     1.15.1             pyh9f0ad1d_0    conda-forge
botocore                  1.18.1             pyh9f0ad1d_0    conda-forge
brotlipy                  0.7.0           py38h4d0b108_1000    conda-forge
ca-certificates           2020.7.22                     0  
certifi                   2020.6.20                py38_0  
cffi                      1.14.3           py38hed5b41f_0  
click                     7.1.2                      py_0  
cryptography              3.1              py38h52adbb4_0    conda-forge
decorator                 4.4.2                      py_0  
idna                      2.10               pyh9f0ad1d_0    conda-forge
intel-openmp              2019.4                      233  
ipykernel                 5.3.4            py38h5ca1d4c_0  
ipython                   7.18.1           py38h5ca1d4c_0  
ipython_genutils          0.2.0                    py38_0  
jedi                      0.17.2                   py38_0  
jmespath                  0.10.0             pyh9f0ad1d_0    conda-forge
joblib                    0.16.0                     py_0  
jupyter_client            6.1.6                      py_0  
jupyter_core              4.6.3                    py38_0  
libcxx                    10.0.0                        1  
libedit                   3.1.20191231         h1de35cc_1  
libffi                    3.3                  hb1e8313_2  
libgfortran               3.0.1                h93005f0_2  
libsodium                 1.0.18               h1de35cc_0  
llvm-openmp               10.0.0               h28b9765_0  
mkl                       2019.4                      233  
mkl-service               2.3.0            py38hfbe908c_0  
mkl_fft                   1.2.0            py38hc64f4ea_0  
mkl_random                1.1.1            py38h959d312_0  
ncurses                   6.2                  h0a44026_1  
numpy                     1.19.1           py38h3b9f5b6_0  
numpy-base                1.19.1           py38hcfb5961_0  
openssl                   1.1.1h               haf1e3a3_0  
pandas                    1.1.1            py38hb1e8313_0  
parso                     0.7.0                      py_0  
pexpect                   4.8.0                    py38_1  
pickleshare               0.7.5                 py38_1001  
pip                       20.2.2                   py38_0  
prompt-toolkit            3.0.7                      py_0  
ptyprocess                0.6.0                    py38_0  
pycparser                 2.20               pyh9f0ad1d_2    conda-forge
pygments                  2.7.1                      py_0  
pyopenssl                 19.1.0                     py_1    conda-forge
pysocks                   1.7.1            py38h32f6830_1    conda-forge
python                    3.8.5                h26836e1_1  
python-dateutil           2.8.1                      py_0  
python-dotenv             0.14.0                   pypi_0    pypi
python_abi                3.8                      1_cp38    conda-forge
pytz                      2020.1                     py_0  
pyzmq                     19.0.2           py38hb1e8313_1  
readline                  8.0                  h1de35cc_0  
s3transfer                0.3.3            py38h32f6830_1    conda-forge
scikit-learn              0.23.2           py38h959d312_0  
scipy                     1.5.0            py38hbab996c_0  
setuptools                49.6.0                   py38_0  
six                       1.15.0                     py_0  
sqlite                    3.33.0               hffcf06c_0  
threadpoolctl             2.1.0              pyh5ca1d4c_0  
tk                        8.6.10               hb0a8c7a_0  
tornado                   6.0.4            py38h1de35cc_1  
traitlets                 4.3.3                    py38_0  
urllib3                   1.25.10                    py_0    conda-forge
wcwidth                   0.2.5                      py_0  
wheel                     0.35.1                     py_0  
xz                        5.2.5                h1de35cc_0  
zeromq                    4.3.2                hb1e8313_3  
zlib                      1.2.11               h1de35cc_3  
